requirements.txt
transformers
torch
requests
Pillow 
uvicorn
streamlit
python-multipart